Move ParallProofs class to expressions
Ref. #625
